backend/requirements.txt
accelerate==0.25.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
Authlib==1.3.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bentoml==1.1.11
better-exceptions==0.3.3
bitsandbytes==0.41.3.post2
blinker==1.7.0
blis==1.3.0
boto3==1.34.44
botocore==1.34.44
build==0.10.0
cachetools==5.3.2
cassandra-driver==3.29.0
catalogue==2.0.10
cattrs==23.1.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
circus==0.18.0
click==8.1.7
click-option-group==0.5.6
cloudpathlib==0.21.1
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contextlib2==21.6.0
cryptography==42.0.3
cuda-python==12.3.0
cymem==2.0.11
dataclasses-json==0.6.3
dataclasses-json-speakeasy==0.5.11
datasets==2.17.0
deepmerge==1.1.1
Deprecated==1.2.14
dill==0.3.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
Django==5.0.2
django-cors-headers==4.3.1
django-grappelli==4.0.2
django-light==0.1.0.post3
django-model-info==2024.11.5
django-simpleui==2025.5.262
django-yugabytedb==4.0.0.post1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
docker-pycreds==0.4.0
docstring-parser==0.15
docx==0.2.4
einops==0.7.0
emoji==2.10.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
errant==3.0.0
exception==0.1.0
faiss-cpu==1.7.4
fast-rest-framework==0.0.1
fastapi==0.109.2
fastcore==1.5.29
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.0
fs==2.4.16
fsspec==2023.10.0
fuzzywuzzy==0.18.0
geomet==0.2.1.post1
ghapi==1.0.4
gingerit==0.0.0.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.28.0
googleapis-common-protos==1.62.0
gramformer @ git+https://github.com/PrithivirajDamodaran/Gramformer.git@23425cd2e98a919384cab6156af8adf1c9d0639a
greenlet==3.2.0
grpcio==1.60.1
grpcio-health-checking==1.60.1
grpcio-tools==1.60.1
gunicorn==23.0.0
h11==0.14.0
housekeeping==1.1
httpcore==1.0.3
httptools==0.6.1
httpx==0.28.1
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
InstructorEmbedding==1.0.1
Jinja2==3.1.2
jiter==0.10.0
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kubernetes==29.0.0
langchain==0.0.347
langchain-community==0.0.24
langchain-core==0.1.28
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.1.10
language_data==1.3.0
language_tool_python==2.9.4
Levenshtein==0.27.1
llama_cpp_python==0.2.20
llvmlite==0.44.0
lxml==5.1.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
murmurhash==1.0.13
mypy-extensions==1.0.0
networkx==3.5
nltk==3.8.1
numba==0.61.2
numpy==2.2.6
nvidia-ml-py==11.525.150
oauthlib==3.2.2
ollama==0.5.1
onnxruntime==1.17.0
openai==0.28.0
openai-whisper==20240930
openllm==0.4.44
openllm-client==0.4.44
openllm-core==0.4.44
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
optimum==1.17.1
orjson==3.9.14
overrides==7.7.0
packaging==23.2
pandas==2.1.3
pathspec==0.12.1
pbr==6.0.0
pdfminer.six==20231228
peft==0.8.2
pgvector==0.2.5
Pillow==10.1.0
pip-requirements-parser==32.0.1
pip-tools==7.3.0
platformdirs==4.1.0
playwright==1.51.0
posthog==3.4.1
preshed==3.0.10
prometheus_client==0.20.0
protobuf==4.25.1
psutil==5.9.6
psycopg==3.1.18
psycopg2==2.9.9
pulsar-client==3.4.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.8.1b0
pydiffx==1.1
pydot==4.0.0
pyee==12.1.1
Pygments==2.17.2
PyJWT==2.9.0
pyparsing==3.1.1
pypdf==3.17.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-json-logger==2.0.7
python-Levenshtein==0.27.1
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
RapidFuzz==3.13.0
RBTools==5.0
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.1
schema==0.7.5
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.40.6
setproctitle==1.3.3
shtab==1.7.0
simple-di==0.1.5
six==1.16.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
srsly==2.5.1
starlette==0.36.3
stevedore==5.1.0
streamlit==1.29.0
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
texttable==1.7.0
thinc==8.3.6
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.1
torchaudio==2.1.1+cpu
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
transformers==4.37.2
trl==0.7.11
typer==0.9.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tyro==0.7.3
tzdata==2023.3
tzlocal==5.2
unstructured==0.12.4
unstructured-client==0.18.0
urllib3==2.0.7
uvicorn==0.27.1
validators==0.22.0
virtualenv==20.25.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.0.0
virtualenvwrapper-win==1.2.7
wandb==0.16.3
wasabi==1.1.3
watchdog==3.0.0
watchfiles==0.21.0
weasel==0.4.1
weaviate-client==4.4.4
websocket-client==1.7.0
websockets==12.0
whitenoise==6.9.0
wrapt==1.16.0
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
